API-аудит: сверка публичного контракта с поведением прода
Журнал аудита пополнился событием `user_2fa_disabled`.

У отправки сообщений появился месячный предел — 30 000 сообщений в один чат за
календарный месяц, с прогрессивной паузой при превышении. Описан на странице
лимитов вместе с отдельным пределом для выборки списка чатов номером страницы.

Ответ `409` описан у методов, где он возможен, и называет конфликтующее поле
в `errors[].key`.

Формы: у радиокнопки изначальный выбор задаётся полем `checked`, у пункта
выпадающего списка появилось описание, массив `options` отмечен обязательным,
предел длины `value` — 255 символов, у напоминания проставлены границы
приоритета.

Кнопки сообщения: предел длины `url`, правило «ровно одно из `url` и `data`»,
пределы количества приведены к тому, что проверяет API.

Файлы: сроки жизни ссылок, размер 64-битным числом, поведение фильтра
расширений в форме, разделы про ссылку на скачивание и удаление вложений.
Адрес вебхука экспорта должен быть доступен снаружи.

Разметка: заголовок распознаётся и становится жирным; в `content` исходящего
вебхука приходят цитаты и списки. Доставка исходящих вебхуков описана по
фактическому поведению.

Версии: n8n 2.1.1, CLI 2026.8.3, SDK 1.4.0.

+++ sdk/kotlin/examples/upload.kt
@@ -72,7 +72,7 @@ fun main() = runBlocking {
                         key = key,
                         name = filename,
                         fileType = FileType.FILE,
-                        size = fileSize,
+                        size = fileSize.toLong(),
                     )
                 ),
             )
